Refactor ParameterizedArrayFactory
